requirements.txt
langchain
langchain-anthropic
langchain-openai
langchain-community
python-dotenv
pydantic
wikipedia